created the peripheral type
At the basic level, the Gateway and Processor concepts were rather similar. They are both components with identifiers which may, or may not generate emersyx events and which receive events (either from other components or the emersyx code). As a result, I created the Peripheral type which is a supertype for both Gateway and Processor types. In addition, the Receptor type is now standalone and can be implemented or not, on a case-by-case basis.