move irc and telegram apis into their own packages
In the case of the Telegram APIs, several types have generic names such as "User", "Update" or "Message". Since these are the names used in the official Telegram Bot API documentation, we want to keep them in emersyx as well. However, this might lead to problems in the future if any other service provides types with the same names. Some namespacing is required in this case, and in golang this means different packages.