refactoring to simplify the api
This commit is the result of code refactoring aimed to simplify the emersyx API for peripherals and the router. Instead of having xOptions interfaces with methods which return functions, we now have xOptions structs with members. The reason for the switch is the result of consolidation of options which are mandatory for all Peripheral and Router instances. Therefore it makes no sense to have the possibility of "optional" configuration options. Therefore it makes no sense to have fancy methods returning functions, which lead to boilerplate code.