remove PeripheralOptions.Validate() method
The PeripheralOptions.Validate() method was found to be too strict, especially in a testing environment where a log writer and core cannod be provided to the peripheral. As such, they were removed an the only thing remaining was the check on the PeripheralOptions.Identifier member. The complete function is removed altogether.